+++ src/Experimental/Graphics/LineBatch.cpp
@@ -450,8 +450,8 @@ void LineBatch::DrawSphere(
         return;
     }
 
-    const auto rings = std::max(static_cast<int>(segments), 4);
-    const auto sectors = std::max(static_cast<int>(segments), 4);
+    const auto rings = std::max(segments, 4);
+    const auto sectors = std::max(segments, 4);
 
     POMDOG_ASSERT(sectors > 0);
     POMDOG_ASSERT(rings > 0);
